Group and friends APIs added

+++ src/main/java/dev/pawan/sharemate/model/Balance.java
@@ -1,9 +1,18 @@
 package dev.pawan.sharemate.model;
 
-import jakarta.persistence.*;
-import lombok.*;
-
 import java.math.BigDecimal;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "BALANCES")
@@ -24,5 +33,5 @@ public class Balance {
     private User friend;
 
     @Column(nullable = false)
-    private Double amount;
+    private BigDecimal amount;
 }
